refactor: extract updateRack into layout/rack-actions domain module (#1077)
Completes the phase 2 rack/rack-group domain extraction. The bulk of the
rack and rack-group domain modules (layout/rack-actions.ts,
layout/rack-groups.ts) landed alongside phase 3 in PR #1465; updateRack
was the last rack domain logic still inlined in the facade.

- Move updateRack (bayed height rejection, view handling, bayed-group
  numbering batch logic) into layout/rack-actions.ts
- Inject the recorded actions (updateRackRecorded,
  updateRacksBatchRecorded) as parameters, following the existing
  removeBayFromGroup/setBayCount DI precedent, so rack-actions does not
  import command-adapters (which imports rack-actions)
- Facade keeps the same public API; updateRack is now a thin delegation
  like the other rack actions
