+++ src/lib/stores/layout/rack-actions.ts
@@ -21,8 +21,23 @@ import {
   type RackLifecycleCommandStore,
 } from "../commands";
 import type { LayoutStateAccess } from "./types";
-import { getRackGroupCommandAdapter } from "./rack-groups";
+import { getRackGroupCommandAdapter, getRackGroupForRack } from "./rack-groups";
 import { setLayoutNamesRaw } from "./mutators";
+
+/** Recorded single-rack update action injected by the facade. */
+export type UpdateRackRecordedFn = (
+  rackId: string,
+  updates: Partial<Omit<Rack, "devices" | "view">>,
+) => void;
+
+/** Recorded batch rack update action injected by the facade. */
+export type UpdateRacksBatchRecordedFn = (
+  targets: {
+    rackId: string;
+    updates: Partial<Omit<Rack, "devices" | "view">>;
+  }[],
+  description: string,
+) => void;
 
 // =============================================================================
 // Raw Mutators (for undo/redo system — bypass history)
@@ -536,4 +551,88 @@ export function duplicateRack(
   ctx.markDirty();
 
   return { rack: duplicatedRack };
+}
+
+/**
+ * Update a rack's properties
+ * Uses undo/redo support via the injected recorded actions (except for view
+ * changes, which apply directly without history).
+ * @param ctx - Layout state access
+ * @param id - Rack ID to update
+ * @param updates - Properties to update
+ * @param updateRackRecordedFn - Recorded single-rack update (undo/redo)
+ * @param updateRacksBatchRecordedFn - Recorded batch update (undo/redo)
+ */
+export function updateRack(
+  ctx: LayoutStateAccess,
+  id: string,
+  updates: Partial<Rack>,
+  updateRackRecordedFn: UpdateRackRecordedFn,
+  updateRacksBatchRecordedFn: UpdateRacksBatchRecordedFn,
+): void {
+  const rackIndex = ctx.findRackIndex(id);
+  if (rackIndex === -1) return;
+
+  // Check if height change on bayed rack
+  if (updates.height !== undefined) {
+    const group = getRackGroupForRack(ctx, id);
+    if (group?.layout_preset === "bayed") {
+      layoutDebug.state(
+        "updateRack: rejected height change for bayed rack %s",
+        id,
+      );
+      // Silently reject - UI should show toast
+      return;
+    }
+  }
+
+  // Handle view separately (doesn't need undo/redo)
+  if (updates.view !== undefined) {
+    const layout = ctx.getLayout();
+    ctx.setLayout({
+      ...layout,
+      racks: layout.racks.map((r, i) =>
+        i === rackIndex ? { ...r, view: updates.view } : r,
+      ),
+    });
+    ctx.markDirty();
+  }
+
+  // For other properties, use recorded version for undo/redo support
+  const { view: _view, devices: _devices, ...recordableUpdates } = updates;
+  if (Object.keys(recordableUpdates).length === 0) return;
+
+  // BayedRackView renders one shared U-label column read from racks[0], so
+  // all bays must agree on desc_units / starting_unit. When the change
+  // touches those keys on a member of a bayed group, fold the origin and
+  // every diverging peer into a single batch — one undo reverts the whole
+  // group together (#1520).
+  const numberingKeys = ["desc_units", "starting_unit"] as const;
+  const numberingUpdates: Partial<Omit<Rack, "devices" | "view">> = {};
+  for (const key of numberingKeys) {
+    if (key in recordableUpdates) {
+      numberingUpdates[key] = recordableUpdates[key] as never;
+    }
+  }
+
+  const group =
+    Object.keys(numberingUpdates).length > 0
+      ? getRackGroupForRack(ctx, id)
+      : undefined;
+
+  if (group?.layout_preset === "bayed" && group.rack_ids.length > 1) {
+    // Origin gets the full update; peers only get the numbering keys.
+    const targets: {
+      rackId: string;
+      updates: Partial<Omit<Rack, "devices" | "view">>;
+    }[] = [{ rackId: id, updates: recordableUpdates }];
+    for (const peerId of group.rack_ids) {
+      if (peerId === id) continue;
+      targets.push({ rackId: peerId, updates: numberingUpdates });
+    }
+    updateRacksBatchRecordedFn(targets, "Update bayed rack");
+    return;
+  }
+
+  updateRackRecordedFn(id, recordableUpdates);
 }
